[will-squash] adds decisionlib hook
