+++ automation/taskcluster/decisionlib/common.py
@@ -1,0 +1,1 @@
+SlugId = str
